[SQ Smells] Refactor check_ord_tls from peer module to ord module
