+++ tests/fabric/test_peer.py
@@ -2,33 +2,8 @@ from copy import deepcopy
 from unittest import mock
 from unittest.mock import call
 
-from nephos.fabric.peer import check_ord_tls, check_peer, setup_peer, peer_channel_suffix, create_channel
+from nephos.fabric.peer import check_peer, setup_peer, peer_channel_suffix, create_channel
 from nephos.helpers.helm import HelmPreserve
-
-
-class TestCheckOrdTls:
-    OPTS = {
-        "msps": {"ord_MSP": {"namespace": "orderer-namespace"}},
-        "orderers": {"names": ["an-ord"], "msp": "ord_MSP"},
-    }
-
-    @mock.patch("nephos.fabric.peer.execute")
-    def test_check_ord_tls(self, mock_execute):
-        mock_execute.side_effect = [("value", None)]
-        check_ord_tls(self.OPTS)
-        mock_execute.assert_called_once_with(
-            'kubectl get cm -n orderer-namespace an-ord-hlf-ord--ord -o jsonpath="{.data.ORDERER_GENERAL_TLS_ENABLED}"',
-            verbose=False,
-        )
-
-    @mock.patch("nephos.fabric.peer.execute")
-    def test_check_ord_tls_verbose(self, mock_execute):
-        mock_execute.side_effect = [("value", None)]
-        check_ord_tls(self.OPTS, verbose=True)
-        mock_execute.assert_called_once_with(
-            'kubectl get cm -n orderer-namespace an-ord-hlf-ord--ord -o jsonpath="{.data.ORDERER_GENERAL_TLS_ENABLED}"',
-            verbose=True,
-        )
 
 
 class TestCheckPeer:
